fix: Quick fix for "requirement failed: Session isn't active." error
The Livy server timed out after 1 hour. An ideal way to fix this is to change the system so we pass a smaller process to the Spark session, which means we may want to do the GET request in the DAG. However, rather than restructuring the code, we simply update the timeout for now.